Move enums to separate module
